split process sampling to its own files
